requirements.txt
3to2==1.1.1
alembic==1.0.8
Babel==2.6.0
blinker==1.4
Click==7.0
coverage==4.5.3
decorator==4.3.2
dominate==2.3.5
elasticsearch==6.3.1
Flask==1.0.2
Flask-Babel==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.4.0
Flask-Moment==0.7.0
Flask-MySQLdb==0.2.0
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.3.2
Flask-WhooshAlchemy==0.56
Flask-WTF==0.14.2
flipflop==1.0
guess-language==0.2
guess-language-spirit==0.5.3
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.8
MarkupSafe==1.1.1
mysqlclient==1.4.2.post1
pbr==5.1.3
pkg-resources==0.0.0
PyJWT==1.7.1
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
python-openid==2.2.5
pytz==2018.9
six==1.12.0
SQLAlchemy==1.3.1
sqlalchemy-migrate==0.12.0
sqlparse==0.3.0
Tempita==0.5.2
urllib3==1.24.1
visitor==0.1.3
Werkzeug==0.14.1
Whoosh==2.7.4
WTForms==2.2.1
